+++ startup/GafferScene/standardAttributes.py
@@ -142,7 +142,7 @@ Gaffer.Metadata.registerValue(
 )
 
 Gaffer.Metadata.registerValue( "attribute:linkedLights", "label", "Linked Lights" )
-Gaffer.Metadata.registerValue( "attribute:linkedLights", "defaultValue", IECore.StringData( "" ) )
+Gaffer.Metadata.registerValue( "attribute:linkedLights", "defaultValue", "defaultLights" )
 Gaffer.Metadata.registerValue(
 	"attribute:linkedLights",
 	"description",
@@ -152,6 +152,7 @@ Gaffer.Metadata.registerValue(
 	all lights that contribute to illumination by default.
 	"""
 )
+Gaffer.Metadata.registerValue( "attribute:linkedLights", "ui:scene:acceptsSetExpression", True )
 
 Gaffer.Metadata.registerValue( "attribute:filteredLights", "label", "Filtered Lights" )
 Gaffer.Metadata.registerValue( "attribute:filteredLights", "defaultValue", IECore.StringData( "" ) )
@@ -165,6 +166,7 @@ Gaffer.Metadata.registerValue(
 	contribute to illumination by default.
 	"""
 )
+Gaffer.Metadata.registerValue( "attribute:filteredLights", "ui:scene:acceptsSetExpression", True )
 
 Gaffer.Metadata.registerValue( "attribute:gaffer:automaticInstancing", "label", "Automatic Instancing" )
 Gaffer.Metadata.registerValue( "attribute:gaffer:automaticInstancing", "defaultValue", IECore.BoolData( True ) )